third_party/blink/web_tests/http/tests/devtools/coverage/gutter-html-expected.txt

Tests the coverage highlight in sources after the recording finishes.

0:   <html>
1:   <head>
2:   <style>
3: + .used {
4: +   color: black;
5: + }
6:   
7: - .notUsed {
8: -   color: red;
9: - }
10:   </style>
11:   <style>
12: - #not-used {
13: -   color: red;
14: - }
15:   
16: + #also-used {
17: +   color: black;
18: + }
19:   </style>
20:   </head>
21:   <body>
22:   <script>
23: + function used1() {
24: +   window.a += 4;
25: + }
26:   
27: - function unused1() {
28: -   window.a += 1;
29: - }
30:   
31: - function unused2() {
32: -   window.a += 0.5;
33: - }
34:   
35: + function used2() {
36: +   window.a += 4;
37: + }
38:   
39: + function performActions() {
40: +   used1();
41: +   used2();
42: + }
43:   </script>
44:   <div class="used">This is text.</div>
45:   <div id="also-used">This is more text.</div>
46:   </body>
47:   
